libflux: split rpc out to rpc.[ch]
