+++ src/common/libflux/flux.h
@@ -10,6 +10,7 @@
 #include "panic.h"
 #include "event.h"
 #include "request.h"
+#include "rpc.h"
 #include "module.h"
 #include "reparent.h"
 #include "info.h"
